orthographic projection matrix

+++ packages/math/src/Vector3.h
@@ -2,10 +2,10 @@
 
 class Vector3 {
 public:
-	double x, y, z;
+	float x, y, z;
 
 	Vector3() : x(0), y(0), z(0) {};
-	Vector3(double x, double y, double z) : x(x), y(y), z(z) {};
+	Vector3(float x, float y, float z) : x(x), y(y), z(z) {};
 };
 
 inline Vector3 operator+(const Vector3& lhs, const Vector3& rhs) {
